refactor the Store protocol into Core and SDK, simplify `queryGenerator()`
