+++ packages/interfaces/src/store.ts
@@ -1,5 +1,5 @@
 import type { IDecodedMessage, IDecoder } from "./message.js";
-import type { IBaseProtocolCore } from "./protocols.js";
+import type { IBaseProtocolCore, IBaseProtocolSDK } from "./protocols.js";
 
 export enum PageDirection {
   BACKWARD = "backward",
@@ -45,7 +45,16 @@ export type StoreQueryOptions = {
   cursor?: Cursor;
 };
 
-export interface IStore extends IBaseProtocolCore {
+export interface IStoreCore extends IBaseProtocolCore {}
+
+export type IStoreSDK = IBaseProtocolSDK & {
+  protocol: IBaseProtocolCore;
+  createCursor(message: IDecodedMessage): Cursor;
+  queryGenerator: <T extends IDecodedMessage>(
+    decoders: IDecoder<T>[],
+    options?: StoreQueryOptions
+  ) => AsyncGenerator<Promise<T | undefined>[]>;
+
   queryWithOrderedCallback: <T extends IDecodedMessage>(
     decoders: IDecoder<T>[],
     callback: (message: T) => Promise<void | boolean> | boolean | void,
@@ -58,8 +67,4 @@ export interface IStore extends IBaseProtocolCore {
     ) => Promise<void | boolean> | boolean | void,
     options?: StoreQueryOptions
   ) => Promise<void>;
-  queryGenerator: <T extends IDecodedMessage>(
-    decoders: IDecoder<T>[],
-    options?: StoreQueryOptions
-  ) => AsyncGenerator<Promise<T | undefined>[]>;
-}
+};
